Update sprintMode.d
adding display method customization for sprint bar - similar to 'Better bars' feature. Sprint bar can be now in 'standard' mode - it will be displayed whenever health bar is displayed, then it can be in 'always on' mode - it will be always on. And finally it can be in 'dynamic update' mode - where it will fade out if its value didn't change and is full, and fade in if player is in fight mode, or value changes anyhow.